Module Mirage_net.Stats

val create : unit -> stats

create () returns a fresh set of zeroed counters

val rx : stats -> int64 -> unit

rx t size records that we received a packet of length size

val tx : stats -> int64 -> unit

tx t size records that we transmitted a packet of length size

val reset : stats -> unit

reset t resets all packet counters in t to 0