Module type Mirage_net.S

A network interface that serves Ethernet frames.

type error#row

The type for network interface errors.

and error = private [>

| Net.error

]

The type for network interface errors.

val pp_error : error Fmt.t

pp_error is the pretty-printer for errors.

include Mirage_device.S

type t

val disconnect : t -> unit Lwt.t

val write : t -> size:int -> (Cstruct.t -> int) -> (unit, error) Stdlib.result Lwt.t

write net ~size fill allocates a buffer of length size, where size must not exceed the interface maximum packet size (mtu plus Ethernet header). The allocated buffer is zeroed and passed to the fill function which returns the payload length, which may not exceed the length of the buffer. When fill returns, a sub buffer is put on the wire: the allocated buffer from index 0 to the returned length.

val listen : t -> header_size:int -> (Cstruct.t -> unit Lwt.t) -> (unit, error) Stdlib.result Lwt.t

listen ~header_size net fn waits for a packet with size at most header_size + mtu on the network device. When a packet is received, an asynchronous task is created in which fn packet is called. The ownership of packet is transferred to fn. The function can be stopped by calling disconnect in the device layer.

val mac : t -> Macaddr.t

mac net is the MAC address of net.

val mtu : t -> int

mtu net is the Maximum Transmission Unit of net. This excludes the Ethernet header.

val get_stats_counters : t -> stats

Obtain the most recent snapshot of the interface statistics.

val reset_stats_counters : t -> unit

Reset the statistics associated with this interface to their defaults.